split transformer
